requirements.txt
aiohttp
aria2p
async_generator
bs4
beautifulsoup4>=4.10.0
covid>=2.5.2
dnspython
glitch_this
fontTools
emoji
gitpython
google-api-python-client
google-auth-oauthlib
googletrans==4.0.0-rc1
google-trans-new
langdetect
google_images_download>=2.8.0
gTTS>=2.2.3
gTTS-token>=1.1.3
hachoir
httplib2
humanize
heroku3
lxml==4.6.3
git+https://github.com/johnwmillr/LyricsGenius.git
lyricsgenius
oauth2client
#opencv-python-headless==4.5.3.56
psycopg2
pybase64>=1.2.1
pylast==4.5.0
Pyrogram==1.4.15
pySmartDL
python-barcode
python-dotenv
youtube-dl>=2021.12.17
youtube-search-python>=1.6.2
youtube-search>=2.1.0
pytz
qrcode
requests>=2.27.1
git+https://github.com/bisoncorps/search-engine-parser
speedtest-cli>=2.1.3
sqlalchemy>=1.2
spotdl
selenium>=4.1.2
telethon==1.24.0
telethon-session-sqlalchemy>=0.2.16
urbandict>=0.6.1
wikipedia>=1.4.0
spotify_token
twitter_scraper
chromedriver-autoinstaller
Pillow
telegraph
lottie
git+https://github.com/Joeclinton1/google-images-download.git
pydub
emoji-country-flag
cairosvg
imagedown
pyaztro
